* changes:
  Change addJumpIfR0{Greater|Less}Than() method to take long argument
  Change add{Div|Mul}() method to take long argument instead of int

+++ src/android/net/apf/ApfV4Generator.java
@@ -357,7 +357,7 @@ public class ApfV4Generator {
             return this;
         }
 
-        Instruction addUnsigned(int imm) {
+        Instruction addUnsigned(long imm) {
             mIntImms.add(IntImmediate.newUnsigned(imm));
             return this;
         }
@@ -770,14 +770,14 @@ public class ApfV4Generator {
     /**
      * Add an instruction to the end of the program to multiply register R0 by {@code value}.
      */
-    public ApfV4Generator addMul(int val) {
+    public ApfV4Generator addMul(long val) {
         return append(new Instruction(Opcodes.MUL).addUnsigned(val));
     }
 
     /**
      * Add an instruction to the end of the program to divide register R0 by {@code value}.
      */
-    public ApfV4Generator addDiv(int val) {
+    public ApfV4Generator addDiv(long val) {
         return append(new Instruction(Opcodes.DIV).addUnsigned(val));
     }
 
@@ -884,7 +884,7 @@ public class ApfV4Generator {
      * Add an instruction to the end of the program to jump to {@code target} if register R0's
      * value is greater than {@code value}.
      */
-    public ApfV4Generator addJumpIfR0GreaterThan(int val, String tgt) {
+    public ApfV4Generator addJumpIfR0GreaterThan(long val, String tgt) {
         return append(new Instruction(Opcodes.JGT).addUnsigned(val).setTargetLabel(tgt));
     }
 
@@ -892,7 +892,7 @@ public class ApfV4Generator {
      * Add an instruction to the end of the program to jump to {@code target} if register R0's
      * value is less than {@code value}.
      */
-    public ApfV4Generator addJumpIfR0LessThan(int val, String tgt) {
+    public ApfV4Generator addJumpIfR0LessThan(long val, String tgt) {
         return append(new Instruction(Opcodes.JLT).addUnsigned(val).setTargetLabel(tgt));
     }
